Update scrape_gm.py
HEADS UP: I didnt see the update piror to updating the code this monring.

I kept code as close to the original code as possible

Fixed a couple things: Popular times at can show while the times still load. Its better to detect what we are looking to parse.

The system works as it did originally with only changes made to the string parser and hour/time convert format.